arreglado eliminar comentarios por usuario creador

+++ controllers/dashboardController.js
@@ -186,10 +186,19 @@ module.exports = {
   },
 
   delete_comment: (req, res, next) => {
-    Match.findByIdAndUpdate(req.params.id, { $pull: { comments: req.params.comment}  })
-    .then(() => Comment.findByIdAndRemove(req.params.comment))
-    .then(res.redirect('/dashboard'))
-      .catch( e=>res.rendirect('/dashboard'));}
+    Comment.findById(req.params.comment, (err, comentario) => {
+      if(comentario._creatorId.toString() !== req.user._id.toString()){
+        res.redirect("/dashboard");
+      }
+      else{
+        Match.findByIdAndUpdate(req.params.id, { $pull: { comments: req.params.comment}  })
+        .then(() => Comment.findByIdAndRemove(req.params.comment))
+        .then(res.redirect('/dashboard'))
+        .catch( e => res.redirect('/dashboard'));
+      }
+    });
+  }
+
 
 
 };
